Changes to UserInfo. Working on editing.
